+++ docs/Whats-New.md
@@ -492,6 +492,7 @@ New:
 - No rearm and reload in EMP or temporal (by CrimRecya)
 - Manually empty and reload ammo (by CrimRecya)
 - No turret unit turn to the target (by CrimRecya & TaranDahl)
+- Sell or undeploy building on impact (by CrimRecya)
 - Unit `Speed` setting now accepts floating point values (by Starkku)
 - Exclusive SuperWeapon Sidebar (by NetsuNegi & CrimRecya)
 - Custom exit cell for infantry factory (by Starkku)